Process launching refactored. Logging refactored.
