+++ src/chat/__init__.py
@@ -22,7 +22,7 @@
 
 # Initialize logging system
 import logging
-logger = logging.getLogger("faf.chat")
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 def user2name(user):
