+++ src/chat/chatter.py
@@ -26,6 +26,7 @@ from chat._avatarWidget import avatarWidget
 
 
 from chat import user2name
+from fa.replay import replay
 import util
 
 import fa
@@ -446,7 +447,7 @@ class Chatter(QtGui.QTableWidgetItem):
     @QtCore.pyqtSlot()
     def viewReplay(self):
         if self.name in client.instance.urls:
-            fa.exe.replay(client.instance.urls[self.name])
+            replay(client.instance.urls[self.name])
 
     @QtCore.pyqtSlot()
     def viewVaultReplay(self):
