+++ src/coop/_coopwidget.py
@@ -21,6 +21,8 @@
 
 
 from PyQt4 import QtCore, QtGui
+import fa
+from fa.replay import replay
 import util
 
 from PyQt4.QtNetwork import QNetworkAccessManager, QNetworkRequest
@@ -28,13 +30,14 @@ from PyQt4.QtNetwork import QNetworkAccessManager, QNetworkRequest
 from games.gameitem import GameItem, GameItemDelegate
 from coop.coopmapitem import CoopMapItem, CoopMapItemDelegate
 from games.hostgamewidget import HostgameWidget
-from games import logger
-from fa import Faction
+from fa import faction
 import random
 import fa
 import modvault
 import os
 
+import logging
+logger = logging.getLogger(__name__)
 
 
 FormClass, BaseClass = util.loadUiType("coop/coop.ui")
@@ -112,7 +115,7 @@ class CoopWidget(FormClass, BaseClass):
         faf_replay.write(reply.readAll())
         faf_replay.flush()
         faf_replay.close()  
-        fa.exe.replay(os.path.join(util.CACHE_DIR, "temp.fafreplay"))    
+        replay(os.path.join(util.CACHE_DIR, "temp.fafreplay"))
         
     def processLeaderBoardInfos(self, message):
         ''' Process leaderboard'''
@@ -214,16 +217,16 @@ class CoopWidget(FormClass, BaseClass):
         if not hasattr(item, "mapUrl") :
             return
         
-        if not fa.exe.available():
+        if not fa.instance.available():
             return
             
         self.client.games.stopSearchRanked()
         self.gamemap = fa.maps.link2name(item.mapUrl)
         
-        fa.exe.checkMap(self.gamemap, force=True)
+        fa.check.checkMap(self.gamemap, force=True)
         
         # A simple Hosting dialog.
-        if fa.exe.check("coop"):     
+        if fa.check.check("coop"):
             hostgamewidget = HostgameWidget(self, item)
             
             if hostgamewidget.exec_() == 1 :
@@ -315,12 +318,12 @@ class CoopWidget(FormClass, BaseClass):
         '''
         Slot that attempts to join a game.
         '''
-        if not fa.exe.available():            
+        if not fa.instance.available():
             return
         
         passw = None 
         
-        if fa.exe.check(item.mod, item.mapname, None, item.mods):
+        if fa.check.check(item.mod, item.mapname, None, item.mods):
             if item.access == "password" : 
                 passw, ok = QtGui.QInputDialog.getText(self.client, "Passworded game" , "Enter password :", QtGui.QLineEdit.Normal, "")
                 if ok:
