+++ src/downloadManager/__init__.py
@@ -24,7 +24,7 @@ import os
 import util
 import warnings
 
-logger= logging.getLogger("faf.downloader")
+logger= logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 VAULT_PREVIEW_ROOT = "http://www.faforever.com/faf/vault/map_previews/small/"
