+++ src/fa/__init__.py
@@ -1,4 +1,4 @@
-#-------------------------------------------------------------------------------
+# -------------------------------------------------------------------------------
 # Copyright (c) 2012 Gael Honorez.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the GNU Public License v3.0
@@ -17,39 +17,36 @@
 #-------------------------------------------------------------------------------
 
 
-
-
-
 # Initialize logging system
-import logging
 from PyQt4 import QtCore
 import os
 import util
-logger= logging.getLogger("faf.fa")
+
+import logging
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 GPGNET_HOST = "faforever.com"
 GPGNET_PORT = 8000
 
 
-class Faction:
-    UEF      = "/uef"
-    CYBRAN   = "/cybran"
-    AEON     = "/aeon"
-    SERAPHIM = "/seraphim"
-
+# We only want one instance of Forged Alliance to run, so we use a singleton here (other modules may wish to connect to its signals so it needs persistence)
+from process import instance as instance
+from play import play as play
 
 
 # This is the game path, a string pointing to the player's actual install of Forged Alliance
 gamepath = None
 
+
 def loadPathSC():
     global gamepathSC
     settings = QtCore.QSettings("ForgedAllianceForever", "FA Lobby")
     settings.beginGroup("SupremeCommanderVanilla")
-    gamepathSC = settings.value("app/path")    
+    gamepathSC = settings.value("app/path")
     settings.endGroup()
-    
+
+
 def savePathSC(path):
     global gamepathSC
     gamepathSC = path
@@ -59,13 +56,15 @@ def savePathSC(path):
     settings.endGroup()
     settings.sync()
 
+
 def loadPath():
     global gamepath
     settings = QtCore.QSettings("ForgedAllianceForever", "FA Lobby")
     settings.beginGroup("ForgedAlliance")
-    gamepath = settings.value("app/path")    
+    gamepath = settings.value("app/path")
     settings.endGroup()
-    
+
+
 def savePath(path):
     global gamepath
     gamepath = path
@@ -74,30 +73,30 @@ def savePath(path):
     settings.setValue("app/path", gamepath)
     settings.endGroup()
     settings.sync()
-       
-       
+
+
 def writeFAPathLua():
-    '''
+    """
     Writes a small lua file to disk that helps the new SupComDataPath.lua find the actual install of the game
-    '''
-    name =  os.path.join(util.APPDATA_DIR, u"fa_path.lua")
+    """
+    name = os.path.join(util.APPDATA_DIR, u"fa_path.lua")
     code = u"fa_path = '" + gamepath.replace(u"\\", u"\\\\") + u"'\n"
+
     if gamepathSC:
         code = code + u"sc_path = '" + gamepathSC.replace(u"\\", u"\\\\") + u"'\n"
-    if (os.path.isfile(name)):
-        os.remove(name)
-    lua = open(name, "w")
+
+    lua = open(name, "w+")
     lua.write(code.encode("utf-8"))
-    lua.flush()         
-    os.fsync(lua.fileno()) # Ensuring the file is absolutely, positively on disk.
+    lua.flush()
+    os.fsync(lua.fileno())  # Ensuring the file is absolutely, positively on disk.
     lua.close()
-        
+
 
 # Initial Housekeeping
 loadPath()
 loadPathSC()
 
-import exe
+import check
 import maps
 import replayserver
 import relayserver
@@ -106,3 +105,4 @@ import updater
 import upnp
 import gwreplayinfo
 import gwgametable
+import faction
