+++ src/fa/faction.py
@@ -1,0 +1,4 @@
+UEF = "/uef"
+CYBRAN = "/cybran"
+AEON = "/aeon"
+SERAPHIM = "/seraphim"
