+++ src/fa/proxies.py
@@ -32,7 +32,7 @@ FAF_PROXY_PORT = 9124
 UNIT16 = 8
 
 class proxies(QtCore.QObject):
-    __logger = logging.getLogger("faf.fa.proxies")
+    __logger = logging.getLogger(__name__)
     __logger.setLevel(logging.DEBUG)
 
     def __init__(self, parent=None):
