+++ src/fa/upnp.py
@@ -28,7 +28,7 @@ Created on Mar 22, 2012
 import logging
 import sys
 import util
-logger= logging.getLogger("faf.fa.upnp")
+logger= logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
