+++ src/featuredmods/__init__.py
@@ -21,7 +21,7 @@
 
 
 import logging
-logger = logging.getLogger("faf.featuredmods")
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
