+++ src/games/_gameswidget.py
@@ -20,22 +20,22 @@
 
 
 
-from PyQt4 import QtCore, QtGui
-import util
+import random
 
+from PyQt4 import QtCore, QtGui
+
+import util
 from games.gameitem import GameItem, GameItemDelegate
 from games.moditem import ModItem, mod_invisible, mods
 from games.hostgamewidget import HostgameWidget
-
 from games._mapSelectWidget import mapSelectWidget
-from games import logger
-from fa import Faction
-import random
+from fa import faction
 import fa
 import modvault
 import notificatation_system as ns
 
-
+import logging
+logger = logging.getLogger(__name__)
 
 RANKED_SEARCH_EXPANSION_TIME = 10000 #milliseconds before search radius expands
 
@@ -157,7 +157,7 @@ class GamesWidget(FormClass, BaseClass):
     def handleMatchmakerInfo(self, message):
         action = message["action"]
         if action == "startSearching":
-            if (not fa.exe.check("matchmaker")):
+            if (not fa.check.check("matchmaker")):
                 logger.error("Can't play ranked without successfully updating Forged Alliance.")
                 return            
 
@@ -339,27 +339,16 @@ class GamesWidget(FormClass, BaseClass):
         self.client.send(dict(command="game_matchmaking", mod="matchmaker", state="askingtostop"))
         #self.stopSearchRanked()
 
-        # if (fa.exe.running()):
-        #     QtGui.QMessageBox.information(None, "ForgedAlliance.exe", "FA is already running.")          
-        #     self.stopSearchingTeamMatchmaker()
-        #     return        
-
-        # if (not fa.exe.check("teammatchmaker")):
-        #     self.stopSearchingTeamMatchmaker()
-        #     logger.error("Can't play matchmaker without successfully updating Forged Alliance.")
-        #     return
-
-
 
 
     def startSearchRanked(self, race):
         self.stopSearchingTeamMatchmaker()
-        if (fa.exe.running()):
+        if fa.instance.running():
             QtGui.QMessageBox.information(None, "ForgedAlliance.exe", "FA is already running.")
             self.stopSearchRanked()
             return
 
-        if (not fa.exe.check("ladder1v1")):
+        if (not fa.check.check("ladder1v1")):
             self.stopSearchRanked()
             logger.error("Can't play ranked without successfully updating Forged Alliance.")
             return
@@ -553,7 +542,7 @@ class GamesWidget(FormClass, BaseClass):
     @QtCore.pyqtSlot(bool)
     def toggleUEF(self, state):
         if (state):
-            self.startSearchRanked(Faction.UEF)
+            self.startSearchRanked(faction.UEF)
             self.disconnectRankedToggles()
             self.rankedAeon.setChecked(False)
             self.rankedCybran.setChecked(False)
@@ -566,7 +555,7 @@ class GamesWidget(FormClass, BaseClass):
     @QtCore.pyqtSlot(bool)
     def toggleAeon(self, state):
         if (state):
-            self.startSearchRanked(Faction.AEON)
+            self.startSearchRanked(faction.AEON)
             self.disconnectRankedToggles()
             self.rankedCybran.setChecked(False)
             self.rankedSeraphim.setChecked(False)
@@ -580,7 +569,7 @@ class GamesWidget(FormClass, BaseClass):
     @QtCore.pyqtSlot(bool)
     def toggleCybran(self, state):
         if (state):
-            self.startSearchRanked(Faction.CYBRAN)
+            self.startSearchRanked(faction.CYBRAN)
             self.disconnectRankedToggles()
             self.rankedAeon.setChecked(False)
             self.rankedSeraphim.setChecked(False)
@@ -594,7 +583,7 @@ class GamesWidget(FormClass, BaseClass):
     @QtCore.pyqtSlot(bool)
     def toggleSeraphim(self, state):
         if (state):
-            self.startSearchRanked(Faction.SERAPHIM)
+            self.startSearchRanked(faction.SERAPHIM)
             self.disconnectRankedToggles()
             self.rankedAeon.setChecked(False)
             self.rankedCybran.setChecked(False)
@@ -609,13 +598,13 @@ class GamesWidget(FormClass, BaseClass):
         if (state):
             faction = random.randint(1,4)
             if faction == 1 :
-                self.startSearchRanked(Faction.UEF)
+                self.startSearchRanked(faction.UEF)
             elif faction == 2 :
-                self.startSearchRanked(Faction.CYBRAN)
+                self.startSearchRanked(faction.CYBRAN)
             elif faction == 3 :
-                self.startSearchRanked(Faction.AEON)
+                self.startSearchRanked(faction.AEON)
             else :
-                self.startSearchRanked(Faction.SERAPHIM)
+                self.startSearchRanked(faction.SERAPHIM)
 
             self.disconnectRankedToggles()
             self.rankedAeon.setChecked(False)
@@ -631,14 +620,14 @@ class GamesWidget(FormClass, BaseClass):
         '''
         Slot that attempts to join a game.
         '''
-        if not fa.exe.available():
+        if not fa.instance.available():
             return
 
         self.stopSearchRanked() #Actually a workaround
 
         passw = None
 
-        if fa.exe.check(item.mod, item.mapname, None, item.mods):
+        if fa.check.check(item.mod, item.mapname, None, item.mods):
             if item.access == "password" :
                 passw, ok = QtGui.QInputDialog.getText(self.client, "Passworded game" , "Enter password :", QtGui.QLineEdit.Normal, "")
                 if ok:
@@ -656,13 +645,13 @@ class GamesWidget(FormClass, BaseClass):
         '''
         Hosting a game event
         '''
-        if not fa.exe.available():
+        if not fa.instance.available():
             return
 
         self.stopSearchRanked()
 
         # A simple Hosting dialog.
-        if fa.exe.check(item.mod):
+        if fa.check.check(item.mod):
             hostgamewidget = HostgameWidget(self, item)
 
             if hostgamewidget.exec_() == 1 :
