+++ src/games/hostgamewidget.py
@@ -28,7 +28,7 @@ from fa import maps
 import util
 
 import logging
-logger = logging.getLogger("faf.hostgamewidget")
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 RANKED_SEARCH_EXPANSION_TIME = 10000 #milliseconds before search radius expands
