+++ src/modvault/utils.py
@@ -33,7 +33,7 @@ import warnings
 import cStringIO
 import zipfile
 
-logger = logging.getLogger("faf.modvault")
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 MODFOLDER = os.path.join(util.PERSONAL_DIR, "My Games", "Gas Powered Games", "Supreme Commander Forged Alliance", "Mods")
