+++ src/mumbleconnector/__init__.py
@@ -23,7 +23,7 @@
 from PyQt4 import QtCore
 import logging
 
-logger = logging.getLogger("faf.mumbleconnector")
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 from _mumbleconnector import mumbleConnector as MumbleConnector
