+++ src/profile/__init__.py
@@ -22,7 +22,7 @@
 
 # Initialize logging system
 import logging
-logger = logging.getLogger("faf.profile")
+logger = logging.getLogger(__name__)
 
 from playerstats import Statpage
