+++ src/replays/__init__.py
@@ -21,7 +21,7 @@
 
 
 import logging
-logger = logging.getLogger("faf.replays")
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 from _replayswidget import ReplaysWidget as Replays
