+++ src/replays/replayitem.py
@@ -232,15 +232,15 @@ class ReplayItem(QtGui.QTreeWidgetItem):
 
                     if "faction" in player :
                        
-                        if player["faction"] == 1 :
+                        if player["faction.py"] == 1 :
                             faction = "UEF"
-                        elif player["faction"] == 2 :
+                        elif player["faction.py"] == 2 :
                             faction = "Aeon"
-                        elif player["faction"] == 3 :
+                        elif player["faction.py"] == 3 :
                             faction = "Cybran"
-                        elif player["faction"] == 4 :
+                        elif player["faction.py"] == 4 :
                             faction = "Seraphim"                            
-                        elif player["faction"] == 5 :
+                        elif player["faction.py"] == 5 :
                             faction = "Nomads"     
                         else :
                             faction = "Broken"
