+++ src/stats/__init__.py
@@ -26,7 +26,7 @@ import logging
 import urllib
 import util
 
-logger = logging.getLogger("faf.ladder")
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 from _statswidget import StatsWidget as Stats
