+++ src/tourneys/__init__.py
@@ -24,7 +24,7 @@ from PyQt4 import QtCore
 from PyQt4 import QtWebKit
 import logging
 
-logger = logging.getLogger("faf.tourneys")
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 from _tournamentswidget import TournamentsWidget as Tourneys
