+++ src/tutorials/__init__.py
@@ -23,7 +23,7 @@
 from PyQt4 import QtCore
 import logging
 
-logger = logging.getLogger("faf.tutorials")
+logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 from _tutorialswidget import tutorialsWidget as Tutorials
